Rename 'example' to 'example_popit_project'.
Having a sample project called 'example' turns out to be a bad idea.
What if there is another module called example on the python path,
and that other module turns up first? Bad things, that's what,
and with an unhelpful error message.

Also remove code to add the parent directory of the project to the
python path as it turns out Django is already doing this.