Fix running firefox browser tests
In firefox it seems that the hover then click combo only works for
elements that are visible in the viewport. When an element is outside
the viewport these tests fail with a
`Selenium::WebDriver::Error::ElementNotVisibleError` when run with the
firefox driver.

This patch uses javascript to scroll the required element into view
before interacting with it, which fixes these failures when I run
against firefox locally. Hopefully it will also fix the travis build.