Fix browser editing bug
*sigh* This ended up being a *much* simpler fix than I started with.
Because we're overriding the `set` method for person and organisation
models, we can't expect backbone to always call it with all the fields,
this was causing some of the data to be lost when saving under certain
circumstances.

This fix simply returns early from `nest` if the attributes don't include
the nested model.