Wait for server to start before running tests
On some setups (notably in a vagrant box) the first test sometimes fails
because the server takes longer than 5 seconds to start. [1]

This change uses netcat (nc(1)) to wait for the server to open port 3100
before it continues running the tests.

Added an echo statement to make it more obvious what's going on if the
test server isn't running for some reason, thanks @mhl for the
suggestion!

[1] https://github.com/mysociety/popit/issues/280#issuecomment-31930573