Make sure images are indexed after adding/removing
Because images are added to the object within popit rather than over the
API we need to manually reindex the instance after adding or removing an
image.

Ideally this code can eventually be removed in favour of just POSTing to
the API directly, but in the meantime this seems like the simplest fix.