Route to apps using the model id instead of slug
Since the slug is not a unique field in the database and it doesn't
appear in the popolo schema we're replacing it with the id field. Users
can still specify a custom id if they'd like to have "pretty" urls, but
when they do that using the id field the uniqueness is enforced.