Fix issue where /about fields default to false
The issue was the all fields were getting cast to String values. When
the page was saved the correct settings were saved in memory, but a
string representation of the settings were saved to the database. So the
problem didn't become apparent until the process was restarted and the
settings were loaded from the database again.