Remove inconsistent image indexing test
This test fails more often than it passes due to elasticsearch's refresh
taking longer than the next request does.

In the future this can potentially be added back in as a faster test
using javascript with mocha/supertest when we have access to
elasticsearch so we can force a refresh.