remove nested model from the collection rather than destroying them
if we destroy them then if the list of models has been fetched from the
server the backbone events do not get called in the correct order and we
end up trying to sync the nested model which does not work as it doesn't
have a url. By removing it from the collection we make sure to trigger
the correct series of events to cause the sync to happen on the parent
model.

This was only a problem on deleting more than one item from a list as
the initial list is loaded from JSON on the page and treated differently
by backbone as it thinks it's new.

Fixes #283