call destroy directly on posts model for deleting
Becuase posts are a separate model and not nested within the
organization we need to destroy them directly rather than removing them
from the nested collection and then rely on that cascading the save up
to the organization model. If we don't do this then no API call is made
and the post isn't deleted on the server.

Fixes #406