Fix inability to save most details on organization page
The organization page was using the same code as the person page to save
edits with the result that only name and summary were being save. So we
add new code for saving organizations along with IDs to the person and
org page so we can attach events to the buttons on the correct pages.

Fixes #456