Fix bug with new organization button
The new organization button wasn't always working. This was because
there was some old code that was trying to open a modal to create the
organization in.

Instead it makes more sense just to link directly to the new entity
pages now (for both people and orgs) because they're no longer in
modals.