Move form into own view and make it behave the same
As before the edit form is now done inline. Unlike before the submit
mechanism, for now, just submits to the server and then reloads the
page. In the future we can make this a bit snazzier, but this works
nicely for now without having to worry about updating page state.