do not save blank entries for other_names etc
Without this clicking add for other_names etc and then not filling in
any details and then clicking save will result in a blank entry in the
other_names field of the person. This then causes #528 to occur when
saving a photo so we remove any entries that are entirely blank.