Fix login page for hosting app
Trying to get the login page working on both the hosting app and the
instance app was proving to be a lot of trouble. For the time being just
move the login to the hosting app. This does break guest logins, but
we're going to have to reconsider how they're going to work now we're
switching to site-wide accounts.