Upgrade user's account after successful login or registration
When a user successfully logs in or registers after being prompted to
create an account create a new permission document which links their
account to the instance being upgraded. This uses the
legacyAccountUpgrade middleware, which looks for a legacyInstanceId in
the session and if found uses it to create appropriate permissions.