Check if user has already upgraded their legacy account
We don't want to allow the user to upgrade their legacy account more
than once since it would result in duplicate entries in the permissions
collection. If they try to login with a legacy account then an error
will be shown telling them to login with their new site-wide account.