Simplify new instance creation process
Previously the user had to enter a slug and email address, then they'd
receive an email with a confirmation link, which would lead to a page
with another button that they needed to click before their instance was
finally created. This has now been replaced with a form with a slug
field, which once completed takes the user directly to their new
instance.