remove entity-save-changes class from person submit button
the enter-to-save augmenter triggers a click on anything with a class of
entity-save-changes when you hit enter in an input box. If the submit
button has this class then it means there are two items with that class
on the page which means that sometimes the form is submitted twice. This
can cause multiple memberships to be added.

Fixes #599