allow instance DB name to be different from slug
when instantiating the db connection for an instance check for a dbname
property on the instance. If present use that as the DB name suffix,
otherwise fall back to the slug name for backwards compatability

Note that as connection details are cached any operation that changes
the db either needs to remove the cache entry or restart the popit
process.

Fixes #583
Part of #516