Add examples for contact details fields
It's completely unclear to most users what should go in the 'Label',
'Value', 'Type' and 'Note' fields should be when adding or editing a
contact detail for a person. There's an argument that we should think
more broadly about how to redesign the entry of these kinds of details,
but in the interim this change (adding examples for three out of four
of those fields) at least provides some help for users until we have
time to do something more involved.  I left out the 'Note' field since
it's unclear to me at the moment how it should be used - I assumed this
was for free-form text for any useful notes about that contact detail,
but the example in the Popolo spec ("e.g. for grouping contact details
by physical location" [1]) makes wonder about that.

 [1] http://popoloproject.com/specs/contact-detail.html

This is an interim fix for #525.