Update the service name mongodb -> mongod to match package upgrade
We recently upgraded our version of MongoDB to the one provided by 10gen
rather than the Debian packaged version. One of the changes in these new
packages was that the service name in the Provides: line of MongoDB has
changed from 'mongodb' to 'mongod'. This means that the call to
update-rc.d for the PopIt sysvinit was emitting these errors:

  update-rc.d: using dependency based boot sequencing
  insserv: Service mongodb has to be enabled to start service popit-staging
  insserv: exiting now!
  update-rc.d: error: insserv rejected the script header

Before merging this we should check whether any of the other known
redeployers of PopIt are using this template for the sysvinit script.