Make sure NODE_ENV correctly set and we're using the right master db
If NODE_ENV is not set then it will fall back to development which means
that on a production server any settings in a production config will be
ignored. To avoid this we require that NODE_ENV be set.

As a further check compare the master db name passed in on the command
line and the one that is used by a node script to make sure they agree
and bail if they don't