Switch to passport-localapikey-update module
The passport-localapikey module is no longer being maintained and the
latest version, with support for putting the API keys in the headers,
hasn't been released to the npm registry.

To work around these issues passport-localapikey-update has been
created, which has the added bonus of being updated to work with the
latest version of passport, which should hopefully reduce the chance of
problems due to differing versions.