Use OpenNorth image proxy for image resizing
Include the OpenNorth proxy and remove ours.

Changes to the image_proxy helper to use the new URL format and also to
correctly encode the URL. Also to pass in full URLs for local images.
Change the remote URL handingling code in templates to pass those to the
proxy as well.

This also works if you set the image proxy path to an external host.

Fixes #238 and also #527