Templates should use same config as other code
Previously templates had a special config variable which came from
lib/middleware/config.js, this mostly matched the normal config object,
but had a couple of subtle differences such as having a
'hosting_site_base_url' property rather than 'hosting_server.base_url'.

This commit changes the config middleware to simply expose the whole
config object to the templates. This makes it easier to figure out which
bit of the config you should be using and also makes it easier to grep
for a config property and see where it's being used.