Make sure each app has its own template instance
If the templates instance is created outside of the configuration block
then the hosting app and the instance app both end up sharing the same
instance of the templates module. Because both of these apps have an
'index.html' view, they end up treading on each other and getting in a
tangle.

This simply moves the template instance creation into the template
configuration block so it's not cached by node's module system.