Make sure all apps use the trust proxy setting
Without this setting enabled the apps get the incorrect 'req.protocol'
and therefore generate http urls even when the connection has come in
over https.

The hosting-app, instance-app and dev-helpers all have the setting set
individually because they are top level express applications. Other
applications are created using the express-inherit module, so we just
inherit the setting from the parent application in that case.