Remove unnecessary routes from person/org apps
The GET /:id routes are already defined in generic_document, so don't
need to be removed then re-added in the person and orgaization apps.

Previously this was necessary because the person and organization apps
also searched for the relevant model, but this is now also done in the
generic document app.