limit the languages we allow the user to use
Only allow the user to use either the default language or the alt
language for either saving or viewing. If they aren't using either of
those then fall back to the default language.

Also, if the language isn't recognized then fall back to the default
langauge too as otherwise the language saving and display code may break
as it relies on us recognizing all the languages in the tranlsation
objects to determine if a key is a translation object