correctly translate arrays for viewing
If it's an array we need to translate each element of the array
individually rather than handing over the whole thing. In order to
do this we need to pass back the id of links etc so we can compare them.
We don't actually want to hang on to this ID though so throw it away
afterwards as we don't use them.