replace-database: You no longer need to supply the old DB name
Previously you had to specify the old Mongo database name; this would
either be the same as the slug (if replace-database had never been used
on this instance before) or the instance slug suffixed with a datetime
(if replace-database had been used before). This commit changes that so
that you don't need to specify the old DB name - it's inferred from the
instance slug regardless of which of those situations the instance is
in.

Fixes #663