Don't run travis tests on node 0.8
Express 4.0 requires node 0.10 or higher. The multer module that we've
switched to for handling multipart forms relies on 0.10 specific object
properties. Therefore we're removing support for running popit on node
0.8.

We're not using node 0.8 and I don't think we really want to imply that
we support that version when we don't explicitly develop against it.