add api version 0.2 while still having 0.1
This works by having a v0.1 branch of the api module with a different
name in package.json which we then install. We need to do this as npm
uses the name of the module as the install path under node_modules and
there isn't a way to override this.

There's also code to map the correct module to the relevant path which
involves a bit of unavoidable duplication in the instantiation process.