Only allow deleting of post based memberships
Any posts based memberships will have been added via the API and as we
don't have any interface for editing them then if a membership has a
post_id we only allow users to delete them. While it's irksome it
prevents various unintented consequences that can arise from editing
post memberships in the UI.

This also fixes the situation where a post based membership does not
have an organization_id and it was being added when the document was
saved. We do this by not creating an organization for the membership if
it's missing and there is a post_id on the membership.

Fixes #728