Upgrade express to 3.19.2
This fixes the security issue with serve-static because this version of
express depends on a newer version of connect, which in turn depends on
a version of serve-static which isn't vulnerable to the open redirect
issue.

It also upgrade popit-api to a version which uses the newer version of
express.