clean up person/organization editing read-mode/edit-mode elements
At some point in the recent past, we switched from having read-mode
and edit-mode elements next to each other, interspersed around the
page, to having all the edit-mode elements in a separate template
and collected at the end of the page, away from the read-mode ones.

But many of the old read-mode and edit-mode elements were left in
the templates, despite never actually being shown to users.

This commit removes all the ones I could find. There might be more,
but it's a start.