Update vagrant provisioning script
Make some changes to the vagrant provisioning script to make it easier
to install gems with bundler.

The reason for upgrading for trusty is that the bundler package is a bit
broken on precise and gives permissions errors as it's an outdated
version. The trusty version of bundler is more up-to-date and works
nicely out of the box without any hoop jumping.

- Upgrade to a trusty64 base box
- Install the bundler package