Start and stop test app server at top level
Mocha allows you to put before and after blocks outside describe blocks,
which makes them run globally. This means that the app server is started
before the tests are run, then shutdown at the end. This makes it easier
to write UI tests without having to worry about starting and stopping
the server in each test.